Make subview clipping and lazy both configurable
